# Fix two performance problems in the Newton assembly.