This fixes two problems in the Newton solver assembly discovered by @gassmoeller. The first problem is that the functions call in the inner most loop the function `this->get_pressure_scaling()`. We now store this value in the beginning and reuse it. The second problem was that the tensor contraction for the symmetrisation was very expensive. We could use a similar trick here, making it a lot cheaper.

The stats before:
```
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      88.6s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| Assemble Stokes system          |         1 |      1.66s |       1.9% |
| Assemble Stokes system Newton   |         4 |      16.8s |        19% |
| Assemble Stokes system Picard   |         3 |      7.36s |       8.3% |
| Assemble Stokes system rhs      |         6 |      5.07s |       5.7% |
| Assemble temperature system     |         7 |      12.5s |        14% |
| Build Stokes preconditioner     |         7 |      19.5s |        22% |
| Solve Stokes system             |         7 |      12.6s |        14% |
| Initialization                  |         1 |       5.7s |       6.4% |
| Postprocessing                  |         1 |      2.55s |       2.9% |
| Setup dof systems               |         1 |     0.887s |         1% |
| Setup initial conditions        |         1 |      1.24s |       1.4% |
+---------------------------------+-----------+------------+------------+
```

stats after storing the pressure normalization:
```
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      75.2s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| Assemble Stokes system          |         1 |      1.41s |       1.9% |
| Assemble Stokes system Newton   |         4 |      10.9s |        15% |
| Assemble Stokes system Picard   |         3 |      4.29s |       5.7% |
| Assemble Stokes system rhs      |         6 |      5.09s |       6.8% |
| Assemble temperature system     |         7 |      11.9s |        16% |
| Build Stokes preconditioner     |         7 |      16.6s |        22% |
| Solve Stokes system             |         7 |      13.1s |        17% |
| Initialization                  |         1 |      5.23s |       6.9% |
| Postprocessing                  |         1 |       2.5s |       3.3% |
| Setup dof systems               |         1 |     0.816s |       1.1% |
| Setup initial conditions        |         1 |      1.01s |       1.3% |
+---------------------------------+-----------+------------+------------+
```
Stats after fixing both:
```
+---------------------------------------------+------------+------------+
| Total wallclock time elapsed since start    |      70.9s |            |
|                                             |            |            |
| Section                         | no. calls |  wall time | % of total |
+---------------------------------+-----------+------------+------------+
| Assemble Stokes system          |         1 |      1.39s |         2% |
| Assemble Stokes system Newton   |         4 |      9.34s |        13% |
| Assemble Stokes system Picard   |         3 |      4.21s |       5.9% |
| Assemble Stokes system rhs      |         6 |       4.6s |       6.5% |
| Assemble temperature system     |         7 |      11.2s |        16% |
| Build Stokes preconditioner     |         7 |        16s |        23% |
| Solve Stokes system             |         7 |      12.4s |        17% |
| Initialization                  |         1 |      5.23s |       7.4% |
| Postprocessing                  |         1 |      2.35s |       3.3% |
| Setup dof systems               |         1 |     0.793s |       1.1% |
| Setup initial conditions        |         1 |      1.02s |       1.4% |
+---------------------------------+-----------+------------+------------+
```