@tjhei 2018-06-22
can we close #2274 with this?