@bangerth 2018-06-22
Nice improvements! I wonder whether we need to re-run the figure in our paper :-)