@gassmoeller 2018-06-22
@MFraters: Is this ready for another review? I think you did not address Timo's comment about moving the call to `this->get_pressure_scaling` out of the loop over all quadrature points?