@tjhei 2018-06-24
yes, that is wolfgang's VM. we have to fix that later today.