@gassmoeller 2018-06-24
But the other tester is still broken. Can we convince ourselves that this is only due to floating point imprecision? Maybe rerun a benchmark with analytical solution?