@MFraters 2018-06-24
I don't think the outcome is significantly different, but the convergence patterns is slightly different (some cases better, some cases worse). I think I will have to turn back optimizations to see where the difference comes from.