@MFraters 2018-06-25
Should we merge this, and do the other optimizations in a different pull request?