@bangerth 2018-06-25
Yes, incremental progress is also progress.